+++ lib/madmin.rb
@@ -7,6 +7,7 @@ require "pagy"
 module Madmin
   autoload :Field, "madmin/field"
   autoload :GeneratorHelpers, "madmin/generator_helpers"
+  autoload :MemberAction, "madmin/member_action"
   autoload :Menu, "madmin/menu"
   autoload :Resource, "madmin/resource"
   autoload :ResourceBuilder, "madmin/resource_builder"